+++ DrinkManagerWeb/Views/Shared/Error.cshtml
@@ -1,4 +1,4 @@
-﻿@model ErrorViewModel
+﻿@model DrinkManagerWeb.Models.ViewModels.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
